Unifying calls to chart renderization (mixin).
